% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMGGUM.R
\name{Bayesplot.BMGGUM}
\alias{Bayesplot.BMGGUM}
\title{BMGGUM bayesian convergence diagnosis plotting function}
\usage{
Bayesplot.BMGGUM(x, pars, plot, inc_warmup = F)
}
\arguments{
\item{x}{BMGGUM returned object}

\item{pars}{Names of plotted parameters. They can be "theta", "alpha", "delta", "tau", "cor", "lambda", or a subset of parameters. See vignette for BMGGUM for more details.}

\item{plot}{Types of plots.They can be "density", "trace", or "autocorrelation".}

\item{inc_warmup}{Whether to include warmup iterations or not when plotting. The default is FALSE.}
}
\value{
Selected plots for selected parameters
}
\description{
This function provides plots including density plots, trace plots, and auto-correlation plots to aid model convergence diagnosis.
}
\examples{
\dontrun{
Bayesplot.BMGGUM(mod, 'alpha', 'density', inc_warmup=T)
Bayesplot.BMGGUM(mod, 'delta', 'trace', inc_warmup=F)
Bayesplot.BMGGUM(mod, 'tau', 'autocorrelation', inc_warmup=T)
mod <- BMGGUM(GGUM.Data, delindex, trait, ind, covariate, option)}
}
