% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bayesplot.R
\name{bayesplot}
\alias{bayesplot}
\title{bayesian convergence diagnosis plotting function}
\usage{
bayesplot(x, pars, plot, inc_warmup = FALSE)
}
\arguments{
\item{x}{returned object}

\item{pars}{Names of plotted parameters. They can be "theta", "alpha", "delta", "tau", "cor", "lambda", or a subset of parameters. See vignette for bmggum for more details.}

\item{plot}{Types of plots.They can be "density", "trace", or "autocorrelation".}

\item{inc_warmup}{Whether to include warmup iterations or not when plotting. The default is FALSE.}
}
\value{
Selected plots for selected parameters
}
\description{
This function provides plots including density plots, trace plots, and auto-correlation plots to aid model convergence diagnosis.
}
