% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmggum-package.R
\docType{package}
\name{bmggum-package}
\alias{bmggum-package}
\title{The 'bmggum' package.}
\description{
Full Bayesian estimation of Multidimensional Generalized Graded Unfolding Model (MGGUM)
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
