% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMGGUM.R
\name{Itemplot.BMGGUM}
\alias{Itemplot.BMGGUM}
\title{BMGGUM item plotting function including observable response categories (ORCs)}
\usage{
Itemplot.BMGGUM(x, items = NULL)
}
\arguments{
\item{x}{BMGGUM returned object}

\item{items}{The items to be plotted. The default is all the items.}
}
\value{
Selected ORC plots for selected items
}
\description{
This function provides item plots including observable response categories plots.
}
\examples{
\dontrun{
Itemplot.BMGGUM(mod, items=1:5)
mod <- BMGGUM(GGUM.Data, delindex, trait, ind, covariate, option)}
}
