% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_itemplot.R
\name{itemplot}
\alias{itemplot}
\title{item plotting function including observable response categories (ORCs)}
\usage{
itemplot(x, items = NULL)
}
\arguments{
\item{x}{returned object}

\item{items}{The items to be plotted. The default is all the items.}
}
\value{
Selected ORC plots for selected items
}
\description{
This function provides item plots including observable response categories plots.
}
\examples{
\donttest{
Data <- c(1,4,4,1,1,1,1,1,1,1,4,1,1,3,1,1,NA,2,NA,3,2,2,2,1,3,2,NA,2,1,1)
Data <- matrix(Data,nrow = 10)
deli <- c(1,-1,2,1,3,-1)
deli <- matrix(deli,nrow = 2)
ind <- c(1,1,2)
ind <- t(ind)
cova <- c(0.70, -1.25, 0.48, -0.47, 0.86, 1.25, 1.17, -1.35, -0.84, -0.55)
mod <- bmggum(GGUM.Data=Data, delindex=deli, trait=2, ind=ind, option=4, covariate=cova)
itemplot(mod, items=1:5)}
}
