% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMGGUM.R
\name{bayesplot.bmggum}
\alias{bayesplot.bmggum}
\title{bmggum bayesian convergence diagnosis plotting function}
\usage{
bayesplot.bmggum(x, pars, plot, inc_warmup = FALSE)
}
\arguments{
\item{x}{bmggum returned object}

\item{pars}{Names of plotted parameters. They can be "theta", "alpha", "delta", "tau", "cor", "lambda", or a subset of parameters. See vignette for bmggum for more details.}

\item{plot}{Types of plots.They can be "density", "trace", or "autocorrelation".}

\item{inc_warmup}{Whether to include warmup iterations or not when plotting. The default is FALSE.}
}
\value{
Selected plots for selected parameters
}
\description{
This function provides plots including density plots, trace plots, and auto-correlation plots to aid model convergence diagnosis.
}
\examples{
\donttest{
Data <- c(1,4,4,1,1,1,1,1,1,1,4,1,1,3,1,1,NA,2,NA,3,2,2,2,1,3,2,NA,2,1,1)
Data <- matrix(Data,nrow = 10)
deli <- c(1,-1,2,1,3,-1)
deli <- matrix(deli,nrow = 2)
ind <- c(1,1,2)
ind <- t(ind)
cova <- c(0.70, -1.25, 0.48, -0.47, 0.86, 1.25, 1.17, -1.35, -0.84, -0.55)
mod <- bmggum(GGUM.Data=Data, delindex=deli, trait=2, ind=ind, option=4, covariate=cova)
bayesplot.bmggum(mod, 'alpha', 'density', inc_warmup=TRUE)
bayesplot.bmggum(mod, 'delta', 'trace', inc_warmup=FALSE)
bayesplot.bmggum(mod, 'tau', 'autocorrelation', inc_warmup=TRUE)}
}
