% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMGGUM.R
\name{bayesplot.bmggum}
\alias{bayesplot.bmggum}
\title{bmggum bayesian convergence diagnosis plotting function}
\usage{
bayesplot.bmggum(x, pars, plot, inc_warmup = F)
}
\arguments{
\item{x}{bmggum returned object}

\item{pars}{Names of plotted parameters. They can be "theta", "alpha", "delta", "tau", "cor", "lambda", or a subset of parameters. See vignette for bmggum for more details.}

\item{plot}{Types of plots.They can be "density", "trace", or "autocorrelation".}

\item{inc_warmup}{Whether to include warmup iterations or not when plotting. The default is FALSE.}
}
\value{
Selected plots for selected parameters
}
\description{
This function provides plots including density plots, trace plots, and auto-correlation plots to aid model convergence diagnosis.
}
\examples{
\dontrun{
bayesplot.bmggum(mod, 'alpha', 'density', inc_warmup=T)
bayesplot.bmggum(mod, 'delta', 'trace', inc_warmup=F)
bayesplot.bmggum(mod, 'tau', 'autocorrelation', inc_warmup=T)
mod <- bmggum(GGUM.Data, delindex, trait, ind, covariate, option)}
}
