% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMGGUM.R
\name{itemplot.bmggum}
\alias{itemplot.bmggum}
\title{bmggum item plotting function including observable response categories (ORCs)}
\usage{
itemplot.bmggum(x, items = NULL)
}
\arguments{
\item{x}{bmggum returned object}

\item{items}{The items to be plotted. The default is all the items.}
}
\value{
Selected ORC plots for selected items
}
\description{
This function provides item plots including observable response categories plots.
}
\examples{
\dontrun{
itemplot.bmggum(mod, items=1:5)
mod <- bmggum(GGUM.Data, delindex, trait, ind, covariate, option)}
}
