% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_itemplot.R
\name{itemplot}
\alias{itemplot}
\title{item plotting function including observable response categories (ORCs)}
\usage{
itemplot(x, items = NULL)
}
\arguments{
\item{x}{returned object}

\item{items}{The items to be plotted. The default is all the items.}
}
\value{
Selected ORC plots for selected items
}
\description{
This function provides item plots including observable response categories plots.
}
\examples{
Data <- c(1,4,2,3)
Data <- matrix(Data,nrow = 2)
deli <- c(1,-1,2,1)
deli <- matrix(deli,nrow = 2)
ind <- c(1,2)
ind <- t(ind)
cova <- c(0.70, -1.25)
mod <- bmggum(GGUM.Data=Data,delindex=deli,trait=2,ind=ind,option=4,covariate=cova,iter=5,chains=1)
itemplot(mod, items=1)
}
